refactor: move calculateCurvePoints to app utils
