+++ apps/balancer-tools/src/app/poolsimulator/(utils)/index.ts
@@ -168,3 +168,24 @@ export function convertGqlToAnalysisData(poolData: PoolQuery): AnalysisData {
       };
   }
 }
+
+export function calculateCurvePoints({
+  balance,
+  start = 0,
+}: {
+  balance?: number;
+  start?: number;
+}) {
+  if (!balance || typeof start == "undefined") return [];
+  const numberOfPoints = 100;
+  const initialValue = balance * 0.001;
+  const stepRatio = Math.pow(balance / initialValue, 1 / (numberOfPoints - 1));
+
+  return [
+    start,
+    ...Array.from(
+      { length: numberOfPoints + 20 },
+      (_, index) => initialValue * stepRatio ** index,
+    ),
+  ];
+}
